Rope: eval to support running given code in a sandboxed env w/ npm
An example would be;

```
import hcl from 'hcl-to-json'

let res = hcl("foo bar=5")

rope.return(res)
```

rope.return can be call asynchronously. rope.return covers success state
of the call and can be called once. If a fail message is required to
send rope.fail can be used with an error object like { message: 'fail' }

If there is an error with the provided code, this will be also handled
as rope.fail.